tinkering & adding fixture
